requirements.txt
streamlit
ultralytics==8.2.0
numpy
pillow
opencv-python-headless
requests